try mix video train

+++ viseg/meta_architecture.py
@@ -931,7 +931,7 @@ class VISeg(MinVIS):
             for i in range(max_scores.shape[0]):
                 if max_scores[i] > 0.3:
                     keep_indexes.append(i)
-                    out_list.append({'pred_logits': [pred_logits[i].cpu()],
+                    out_list.append({'pred_logits': [pred_logits[i].to(torch.float32).cpu()],
                                      'pred_masks': [masks[i]]})
             return keep_indexes
